my/zonefile: treat a second SOA as EOF
some ZFAs (e.g., .name) strangely terminate the zone file with a second
SOA, which is technically invalid per IETF and ICANN specs. add a
workaround for these broken zone files by treating a second SOA
encountered in zonefile_read() the same as EOF.

it's not necessary to actually close the FILE object in zonefile_read()
on EOF. this is performed in zonefile_destroy().