reader: get_block(): increase initial zlib decompression buffer size
after the previous commit, we now correctly handle the case where the
decompression buffer is (initially) too small to hold the decompressed
output. however, we can still increase the initial size of this buffer
in order to reduce the amount of looping and reallocation that has to be
done in that case. so, increase it from 2X to 4X of the data block size.