reader: Introduce 'madvise_random' option
This adds an mtbl_reader_options_set_madvise_random() function. On
readers initialized with this option, libmtbl will call
posix_madvise(..., POSIX_MADV_RANDOM); or madvise(..., MADV_RANDOM); on
the mapping set up for the reader's data blocks. This allows the caller
to influence the kernel's readahead behavior, which may help or hurt
performance depending on workload.

The mapping for the reader's index block is left alone since the index
block will generally benefit from readahead. (And we read it
sequentially anyway, when verifying the checksum.)

On some real world seek-heavy workloads enabling this option drastically
decreases the number of pages that are unnecessarily paged in due to
aggressive readahead.