Add mtbl_reader_get_* functions to retrieve trailer metadata
My goal was to create an in-memory representation of the original data.  To
do this, it's very helpful to know the size of the source data, namely the
key & value bytes, and the number of entries.

I added the other getters so that the "mtbl_info" program wouldn't have to
use private APIs to do its job.