reader: Don't keep an open file descriptor corresponding to the mmap()'d file
Previously, mtbl_reader_init_fd() would dup() the file descriptor passed
by the caller, so that we would always have an open file descriptor
corresponding to the file that we mmap(). It turns out that this is not
actually necessary. Per POSIX:

    The mmap() function shall add an extra reference to the file
    associated with the file descriptor fildes which is not removed by a
    subsequent close() on that file descriptor. This reference shall be
    removed when there are no more mappings to the file.

    (http://pubs.opengroup.org/onlinepubs/9699919799/functions/mmap.html)

So, remove the 'fd' field of struct mtbl_reader, and stop keeping
dup()'d file descriptors in mtbl_reader_init_fd().

This change should increase the number of MTBL files that can be
simultaneously opened (with no special permissions or process limit
changes) beyond the open file descriptor limit.