mtbl_verify: mmap() the file starting from the first byte
Previously we attempted to call mmap() with the 'offset' parameter set
to our data block offset, but this offset is not constrained to be a
multiple of the page size, which is required by mmap(). This caused the
mmap() call to fail on MTBL input that included a foreign prefix.

Instead, always call mmap() with the 'offset' parameter set to zero,
i.e., starting from the beginning of the file (which will include the
foreign prefix if any), and skip over this data when verifying the data
blocks.

This makes the currently failing t/test-verify.sh test (with file
t/test-foreign-prefix.data) now succeed.