reader: Sanitize the index block offset
If the index block offset is larger than the total size of the file (as
in t/test-verify-bad-index-block-offset.data), we can end up reading
beyond the end of the file.

Validate the index block offset in mtbl_reader_init_fd(), and refuse to
load the MTBL file if the index block offset is implausible.