compression: Add _mtbl_compress_lz4, _mtbl_compress_lz4hc, _mtbl_decompress_lz4 functions
This commit adds the compression and decompression helper functions for
the LZ4 and LZ4HC algorithms. Note that LZ4 and LZ4HC share a
decompressor.

The LZ4 format doesn't record the size of the uncompressed block, and
the liblz4 API makes it impossible to tell the difference between
a too small output buffer and malformed input when decompressing.
That is, LZ4_decompress_safe() is documented to return:

    the number of bytes decompressed into destination buffer
    (necessarily <= maxDecompressedSize)

        If destination buffer is not large enough, decoding will stop
        and output an error code (<0).

        If the source stream is detected malformed, the function will
        stop decoding and return a negative result.

So, our LZ4 compressed data blocks will carry a 4-byte header containing
a 32-bit little endian integer specifying the size of the decompressed
block.