compression: Fix _mtbl_compress_snappy()
snappy_compress()'s fourth argument is a parameter-return value.
It needs to be initialized to the size of the third argument (the number
of bytes allocated for the output), and on return it specifies the true
length of the compressed output.

This bug was inadvertently introduced by the refactoring in the
compression code. It is now detected by the new 'test-compression' test.