mtbl_merge: Command-line options for block size and compression type
This commit adds two new command-line parameters to mtbl_merge: "-b",
which specifies the data block size, and "-c", which specifies the
compression algorithm to use.

If omitted, the defaults are the same as in previous versions: 8 KB data
blocks and zlib compression.

We still support specifying the block size parameter via the environment
variable MTBL_MERGE_BLOCK_SIZE, but if both the environment variable and
the -b command-line option are present, we generate an error.