block: Use 64 bit restart offsets when necessary.
If a block encodes more than 4GB of entries (as can happen
in the mtbl index block for large mtbl files), some
restart offsets may point outside of the range of a uint32_t.
Detect this condition, and switch to 64 bit restart offsets.

Detecting this condition on writing is trivial. On reading,
the block's size and number of restarts can be used to infer
whether the block was written using 64 bit restart offsets.

The number of restarts remains a uint32_t.